fix context multiple creation
