+++ uui/services/ContextProvider.tsx
@@ -1,48 +1,59 @@
-import React, { createContext, useContext, useEffect, useRef, useState } from "react";
-import { ApiCallOptions, CommonContexts } from "../types";
-import { ContextProviderProps, LegacyContextProvider } from "./LegacyContextProvider";
+import React, { createContext, useContext, useEffect, useMemo, useState } from "react";
+import { ApiCallOptions, CommonContexts, IHasChildren } from "../types";
+import { LegacyContextProvider } from "./LegacyContextProvider";
 import { LayoutContext } from "./LayoutContext";
 import { ModalContext } from "./ModalContext";
 import { NotificationContext } from "./NotificationContext";
-import { HistoryAdaptedRouter, StubAdaptedRouter } from "./routing";
+import { HistoryAdaptedRouter, IHistory4, StubAdaptedRouter } from "./routing";
 import { AnalyticsContext } from "./AnalyticsContext";
 import { LockContext } from "./LockContext";
 import { ErrorContext } from "./ErrorContext";
 import { ApiContext } from "./ApiContext";
 import { UserSettingsContext } from "./UserSettingsContext";
 import { DndContext, DragGhost } from "./dnd";
-import { uuiSkin } from "./SkinContext";
+import { ISkin, uuiSkin } from "./SkinContext";
 
+export interface ContextProviderProps<TApi, TAppContext> extends IHasChildren {
+    apiServerUrl?: string;
+    gaCode?: string;
+    ampCode?: string;
+    loadAppContext?: (api: TApi) => Promise<TAppContext>;
+    apiDefinition?: (processRequest: (request: string, requestMethod: string) => any) => TApi;
+    onInitCompleted(svc: CommonContexts<TApi, TAppContext>): void;
+    history?: IHistory4;
+    skinContext?: ISkin;
+    enableLegacyContext?: boolean;
+}
 
 export const UuiContext = createContext({} as CommonContexts<any, any>);
 
 export const ContextProvider = <TApi, TAppContext>(props: ContextProviderProps<TApi, TAppContext>) => {
     const [isLoaded, setIsLoaded] = useState(false);
-    const uuiContexts = useRef<CommonContexts<TApi, TAppContext>>(getUuiContexts(props));
+    const uuiContexts = useMemo<CommonContexts<TApi, TAppContext>>(() => getUuiContexts(props), []);
 
     useEffect(() => {
         const loadAppContext = props?.loadAppContext || (() => Promise.resolve({} as TAppContext));
 
-        loadAppContext(uuiContexts.current.api).then(appCtx => {
-            uuiContexts.current.uuiApp = appCtx;
-            props.onInitCompleted(uuiContexts.current);
+        loadAppContext(uuiContexts.api).then(appCtx => {
+            uuiContexts.uuiApp = appCtx;
+            props.onInitCompleted(uuiContexts);
             setIsLoaded(true);
         });
     }, []);
 
     // Workaround to discard all errors on navigation. Need to find a better way. YakovZh
-    (uuiContexts.current.uuiErrors as any).discardError();
-    uuiContexts.current.uuiApi.reset();
-    
+    (uuiContexts.uuiErrors as any).discardError();
+    uuiContexts.uuiApi.reset();
+
 
     const children = isLoaded ? props.children : '';
     const enableLegacyContexts = props.enableLegacyContext ?? true;
 
     return (
-        <UuiContext.Provider value={ uuiContexts.current }>
+        <UuiContext.Provider value={ uuiContexts }>
             { enableLegacyContexts
                 ? (
-                    <LegacyContextProvider { ...props } >
+                    <LegacyContextProvider { ...props } uuiContexts={ uuiContexts } >
                         { children }
                         <DragGhost/>
                     </LegacyContextProvider>
