+++ uui/services/index.ts
@@ -2,7 +2,7 @@ export * from './dnd';
 export * from './ScrollManager';
 export * from './LayoutContext';
 export * from './LegacyContextProvider';
-export * from './UuiContext';
+export * from './ContextProvider';
 export * from './AnalyticsContext';
 export * from './ApiContext';
 export * from './ErrorContext';
